Extract parser session grammar
